modified resolution switch timer logic to verify frame first, added more comments

+++ Test_Scripts_Linux/ResolutionSwitch.py
@@ -9,11 +9,17 @@ import re
 from datetime import date, datetime
 
 ap = argparse.ArgumentParser()
+ap.add_argument("-d","--debug", type=bool, default=False, help="Set to True to disable msgs to terminal")
 ap.add_argument("-f","--frame", type=bool, default=False, help="Set to True to enable live view")
+ap.add_argument("-p","--power", type=bool, default=False, help="Set to true when running on the Jenkins server")
 ap.add_argument("-t","--test", type=str, default="res_switch.json", help="Specify .json file to load test cases")
+ap.add_argument("-v","--video", type=str, default="Jabra PanaCast 50", help="Specify which camera to test")
 args = vars(ap.parse_args())
+debug = args["debug"]
 live_view = args["frame"]
 test_file = args["test"]
+power_cycle = args["power"]
+device_name = args["video"]
 
 input_tests = open(test_file)
 json_str = input_tests.read()
@@ -21,7 +27,6 @@ test_cases = json.loads(json_str)
 
 global cap
 global reboots
-debug = True
 current = date.today()
 path = os.getcwd()
 device_num = 0
@@ -46,9 +51,15 @@ log_print("\n{}\n".format(timestamp))
 
 def reboot_device():
     print("Panacast device error")
-    os.system("sudo adb kill-server")
-    os.system("sudo adb devices")
-    os.system("adb reboot")
+    if power_cycle is True:
+        subprocess.check_call(['./power_switch.sh', '0', '0'])
+        time.sleep(3)
+        subprocess.check_call(['./power_switch.sh', '0', '0'])
+    else:
+        os.system("sudo adb kill-server")
+        os.system("sudo adb devices")
+        os.system("adb reboot")
+
     print("Rebooting...")
     time.sleep(55)
     global device_num
@@ -57,7 +68,7 @@ def reboot_device():
 
     # grab reenumerated device
     while True:       
-        device = subprocess.check_output('v4l2-ctl --list-devices 2>/dev/null | grep "Jabra PanaCast 50" -A 1 | grep video', shell=True)
+        device = subprocess.check_output('v4l2-ctl --list-devices 2>/dev/null | grep "{}" -A 1 | grep video'.format(device_name), shell=True)
         device = device.decode("utf-8")
         device_num = int(re.search(r'\d+', device).group())
         cap = cv2.VideoCapture(device_num)
@@ -69,8 +80,10 @@ def reboot_device():
 
 def test_fps(width, height, target_res, frame_count):
     start_frame, test_frame, total_frame, drop_frame = (0 for x in range(4))
+    # set initial resolution
     cap.set(cv2.CAP_PROP_FRAME_WIDTH, width)
     cap.set(cv2.CAP_PROP_FRAME_HEIGHT, height)
+    retval, frame = cap.read()
     switch_start = time.time()
 
     for target in target_res:
@@ -79,14 +92,21 @@ def test_fps(width, height, target_res, frame_count):
         else:
             cap.set(cv2.CAP_PROP_FRAME_WIDTH, target[0])
             cap.set(cv2.CAP_PROP_FRAME_HEIGHT, target[1])
-            switch_end = time.time()
-            switch_time = switch_end - switch_start
+            # check when frame is available to grab after switching resolutions
+            while True:
+                retval, frame = cap.read()
+                w = int(cap.get(cv2.CAP_PROP_FRAME_WIDTH))
+                h = int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT))
+                if retval is True and w == target[0] and h == target[1]:
+                    switch_end = time.time()
+                    switch_time = switch_end - switch_start
+                    break
 
         log_print(55*"=")
         log_print("{}x{} -> {}x{}".format(width, height, target[0], target[1]))
         log_print("Time to switch (ms):   {}\n".format(switch_time * 1000))
         test_start, test_time = (time.time() for x in range(2))
-
+        # grab frames for 30 seconds (900 frames)
         for i in range(0, frame_count):
             retval, frame = cap.read()
             if retval is False:
@@ -122,7 +142,7 @@ def test_fps(width, height, target_res, frame_count):
 
 if __name__ == "__main__":
     # set up camera stream        
-    device = subprocess.check_output('v4l2-ctl --list-devices 2>/dev/null | grep "Jabra PanaCast 50" -A 1 | grep video', shell=True)
+    device = subprocess.check_output('v4l2-ctl --list-devices 2>/dev/null | grep "{}" -A 1 | grep video'.format(device_name), shell=True)
     device = device.decode("utf-8")
     device_num = int(re.search(r'\d+', device).group())
     cap = cv2.VideoCapture(device_num)
@@ -132,7 +152,7 @@ if __name__ == "__main__":
         sys.exit(1)
     if cap.isOpened():
         log_print("PanaCast device found:  {}\n".format(device_num))
-
+    # cylce through all test cases provided by json file
     frame_count = 900
     for fmt in test_cases:
         codec = test_cases[fmt]
